simulator2d/requirements.txt
Cython==0.29.6
numpy==1.16.2
opencv-python==4.0.0.21
PyYAML==5.1
